Add drawer item changes
