+++ app/src/main/java/com/jerboa/ui/components/home/HomeActivity.kt
@@ -2,7 +2,6 @@ package com.jerboa.ui.components.home
 
 import android.content.Context
 import android.util.Log
-import androidx.activity.compose.BackHandler
 import androidx.activity.compose.ReportDrawn
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -43,7 +42,6 @@ import androidx.navigation.NavController
 import com.jerboa.R
 import com.jerboa.VoteType
 import com.jerboa.api.ApiState
-import com.jerboa.closeDrawer
 import com.jerboa.datatypes.types.BlockCommunity
 import com.jerboa.datatypes.types.BlockPerson
 import com.jerboa.datatypes.types.CreatePostLike
@@ -54,8 +52,6 @@ import com.jerboa.datatypes.types.Tagline
 import com.jerboa.db.Account
 import com.jerboa.db.AccountViewModel
 import com.jerboa.db.AppSettingsViewModel
-import com.jerboa.fetchHomePosts
-import com.jerboa.fetchInitialData
 import com.jerboa.isLoading
 import com.jerboa.isRefreshing
 import com.jerboa.loginFirstToast
@@ -73,19 +69,15 @@ import com.jerboa.ui.components.common.getCurrentAccount
 import com.jerboa.ui.components.common.getPostViewMode
 import com.jerboa.ui.components.common.rootChannel
 import com.jerboa.ui.components.common.toCommunity
-import com.jerboa.ui.components.common.toCommunityList
 import com.jerboa.ui.components.common.toCreatePost
-import com.jerboa.ui.components.common.toInbox
 import com.jerboa.ui.components.common.toPost
 import com.jerboa.ui.components.common.toPostEdit
 import com.jerboa.ui.components.common.toPostReport
 import com.jerboa.ui.components.common.toProfile
-import com.jerboa.ui.components.common.toSettings
 import com.jerboa.ui.components.post.PostListings
 import com.jerboa.ui.components.post.edit.PostEditReturn
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.toImmutableList
-import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 
 @OptIn(ExperimentalMaterial3Api::class, ExperimentalComposeUiApi::class)
@@ -219,10 +211,13 @@ fun MainPostListingsContent(
 
     Box(modifier = Modifier.pullRefresh(pullRefreshState)) {
         // zIndex needed bc some elements of a post get drawn above it.
-        PullRefreshIndicator(homeViewModel.postsRes.isRefreshing(), pullRefreshState,
+        PullRefreshIndicator(
+            homeViewModel.postsRes.isRefreshing(),
+            pullRefreshState,
             Modifier
                 .align(Alignment.TopCenter)
-                .zIndex(100f))
+                .zIndex(100f),
+        )
         // Can't be in ApiState.Loading, because of infinite scrolling
         if (homeViewModel.postsRes.isLoading()) {
             LoadingBar(padding = padding)
@@ -350,123 +345,6 @@ fun MainPostListingsContent(
             blurNSFW = blurNSFW,
         )
     }
-}
-
-@Composable
-fun MainDrawer(
-    siteViewModel: SiteViewModel,
-    navController: NavController,
-    accountViewModel: AccountViewModel,
-    homeViewModel: HomeViewModel,
-    scope: CoroutineScope,
-    drawerState: DrawerState,
-    onSelectTab: ((BottomNavTab) -> Unit)?,
-    blurNSFW: Boolean,
-) {
-    val ctx = LocalContext.current
-
-    val accounts = accountViewModel.allAccounts.value
-    val account = getCurrentAccount(accountViewModel)
-
-    BackHandler(drawerState.isOpen) {
-        closeDrawer(scope, drawerState)
-    }
-
-    Drawer(
-        siteRes = siteViewModel.siteRes,
-        unreadCount = siteViewModel.getUnreadCountTotal(),
-        accountViewModel = accountViewModel,
-        navController = navController,
-        isOpen = drawerState.isOpen,
-        onSwitchAccountClick = { acct ->
-            accountViewModel.removeCurrent()
-            accountViewModel.setCurrent(acct.id)
-
-            fetchInitialData(
-                account = acct,
-                siteViewModel = siteViewModel,
-            )
-            fetchHomePosts(
-                account = acct,
-                homeViewModel = homeViewModel,
-            )
-
-            closeDrawer(scope, drawerState)
-        },
-        onSignOutClick = {
-            accounts?.also { accts ->
-                account?.also {
-                    accountViewModel.delete(it)
-                    val updatedList = accts.toMutableList()
-                    updatedList.remove(it)
-
-                    if (updatedList.isNotEmpty()) {
-                        accountViewModel.setCurrent(updatedList[0].id)
-                    }
-                    fetchInitialData(
-                        account = updatedList.getOrNull(0),
-                        siteViewModel = siteViewModel,
-                    )
-                    fetchHomePosts(
-                        account = updatedList.getOrNull(0),
-                        homeViewModel = homeViewModel,
-                    )
-
-                    closeDrawer(scope, drawerState)
-                }
-            }
-        },
-        onClickListingType = { listingType ->
-            homeViewModel.updateListingType(listingType)
-            homeViewModel.resetPosts(account)
-            closeDrawer(scope, drawerState)
-        },
-        onCommunityClick = { community ->
-            navController.toCommunity(id = community.id)
-            closeDrawer(scope, drawerState)
-        },
-        onClickProfile = {
-            onSelectTab?.invoke(BottomNavTab.Profile) ?: run {
-                account?.id?.also {
-                    navController.toProfile(id = it)
-                } ?: run {
-                    loginFirstToast(ctx)
-                }
-            }
-            closeDrawer(scope, drawerState)
-        },
-        onClickSaved = {
-            onSelectTab?.invoke(BottomNavTab.Saved) ?: run {
-                account?.id?.also {
-                    navController.toProfile(id = it, saved = true)
-                } ?: run {
-                    loginFirstToast(ctx)
-                }
-            }
-            closeDrawer(scope, drawerState)
-        },
-        onClickInbox = {
-            onSelectTab?.invoke(BottomNavTab.Inbox) ?: run {
-                account?.also {
-                    navController.toInbox()
-                } ?: run {
-                    loginFirstToast(ctx)
-                }
-            }
-            closeDrawer(scope, drawerState)
-        },
-        onClickSettings = {
-            navController.toSettings()
-            closeDrawer(scope, drawerState)
-        },
-        onClickCommunities = {
-            onSelectTab?.invoke(BottomNavTab.Search) ?: run {
-                navController.toCommunityList()
-            }
-            closeDrawer(scope, drawerState)
-        },
-        blurNSFW = blurNSFW,
-    )
 }
 
 @OptIn(ExperimentalMaterial3Api::class)
