+++ app/src/main/res/values-da/strings.xml
@@ -38,7 +38,6 @@
     <string name="dialogs_top_week">Populært denne uge</string>
     <string name="dialogs_top_month">Populært denne måned</string>
     <string name="dialogs_top_year">Populært i år</string>
-    <string name="dialogs_top_all_time">Populært gennem tiderne</string>
     <string name="dialogs_subscribed">Abonneret</string>
     <string name="dialogs_local">Lokalt</string>
     <string name="dialogs_all">Alle</string>
@@ -70,9 +69,6 @@
     <string name="home_subscribed">Abonneret</string>
     <string name="home_local">Lokalt</string>
     <string name="home_all">Alle</string>
-    <string name="home_saved">Gemt</string>
-    <string name="home_profile">Profil</string>
-    <string name="home_inbox">Indbakke</string>
     <string name="home_settings">Indstillinger</string>
     <string name="home_subscriptions">Abonnementer</string>
     <string name="home_add_account">Tilføj konto</string>
